refactor: split up pkam verb from from verb
